Major GC refactor and generalization.
Push knowledge about almost all specific data types out of the GC
core. We replace the static type table with a pointer to a 'gc_op' struct
as the type tag in the heap. This will use up slightly more space, but is
really much cleaner.

We reimplement the Scheme data types in a new external interface,
scgc.[ch].

# Please enter the commit message for your changes.
# (Comment lines starting with '#' will not be included)
# On branch generalize-gc
# Changes to be committed:
#   (use "git reset HEAD <file>..." to unstage)
#
#	modified:   Makefile
#