Fix .d generation for files in test/
Previously, gcc had only been taking the basename of a file when
generating .d files, meaning that e.g. test/vm.c's dependencies were
attributed to vm.c. We can use gcc's -MT optino to explicitly set the
name of the file to fix this, and eliminate the need for a 'sed' step
in the process.